+++ src/main/java/com/objects/marketbridge/coupon/domain/MemberCoupon.java
@@ -1,6 +1,8 @@
-package com.objects.marketbridge.member.domain;
+package com.objects.marketbridge.coupon.domain;
 
 import com.objects.marketbridge.common.service.port.DateTimeHolder;
+import com.objects.marketbridge.member.domain.BaseEntity;
+import com.objects.marketbridge.member.domain.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
