+++ src/main/java/com/objects/marketbridge/order/service/CreateOrderService.java
@@ -3,9 +3,8 @@ package com.objects.marketbridge.order.service;
 import com.objects.marketbridge.common.service.port.DateTimeHolder;
 import com.objects.marketbridge.member.domain.Address;
 import com.objects.marketbridge.member.domain.Member;
-import com.objects.marketbridge.member.domain.MemberCoupon;
+import com.objects.marketbridge.coupon.domain.MemberCoupon;
 import com.objects.marketbridge.member.service.port.MemberRepository;
-import com.objects.marketbridge.order.domain.CalcTotalDiscountService;
 import com.objects.marketbridge.order.domain.Order;
 import com.objects.marketbridge.order.domain.OrderDetail;
 import com.objects.marketbridge.order.domain.ProductValue;
