+++ src/main/java/com/objects/marketbridge/product/infra/coupon/CouponRepository.java
@@ -1,6 +1,6 @@
 package com.objects.marketbridge.product.infra.coupon;
 
-import com.objects.marketbridge.member.domain.Coupon;
+import com.objects.marketbridge.coupon.domain.Coupon;
 
 import java.util.List;
