+++ src/main/java/com/objects/marketbridge/product/infra/coupon/MemberCouponRepository.java
@@ -1,6 +1,6 @@
 package com.objects.marketbridge.product.infra.coupon;
 
-import com.objects.marketbridge.member.domain.MemberCoupon;
+import com.objects.marketbridge.coupon.domain.MemberCoupon;
 
 import java.util.List;
