+++ src/test/java/com/objects/marketbridge/order/controller/OrderCancelControllerTest.java
@@ -3,9 +3,9 @@ package com.objects.marketbridge.order.controller;
 import com.objects.marketbridge.common.exception.exceptions.CustomLogicException;
 import com.objects.marketbridge.common.interceptor.ApiResponse;
 import com.objects.marketbridge.common.service.port.DateTimeHolder;
-import com.objects.marketbridge.member.domain.Coupon;
+import com.objects.marketbridge.coupon.domain.Coupon;
 import com.objects.marketbridge.member.domain.Member;
-import com.objects.marketbridge.member.domain.MemberCoupon;
+import com.objects.marketbridge.coupon.domain.MemberCoupon;
 import com.objects.marketbridge.member.domain.MembershipType;
 import com.objects.marketbridge.order.controller.dto.ConfirmCancelHttp;
 import com.objects.marketbridge.order.controller.dto.GetCancelDetailHttp;
