+++ src/test/java/com/objects/marketbridge/order/controller/OrderReturnControllerTest.java
@@ -2,9 +2,9 @@ package com.objects.marketbridge.order.controller;
 
 import com.objects.marketbridge.common.exception.exceptions.CustomLogicException;
 import com.objects.marketbridge.common.interceptor.ApiResponse;
-import com.objects.marketbridge.member.domain.Coupon;
+import com.objects.marketbridge.coupon.domain.Coupon;
 import com.objects.marketbridge.member.domain.Member;
-import com.objects.marketbridge.member.domain.MemberCoupon;
+import com.objects.marketbridge.coupon.domain.MemberCoupon;
 import com.objects.marketbridge.member.domain.MembershipType;
 import com.objects.marketbridge.order.controller.dto.ConfirmReturnHttp;
 import com.objects.marketbridge.order.controller.dto.GetReturnDetailHttp;
