+++ src/test/java/com/objects/marketbridge/order/mock/FakeCouponRepository.java
@@ -1,6 +1,6 @@
 package com.objects.marketbridge.order.mock;
 
-import com.objects.marketbridge.member.domain.Coupon;
+import com.objects.marketbridge.coupon.domain.Coupon;
 import com.objects.marketbridge.product.infra.coupon.CouponRepository;
 
 import java.util.List;
