+++ src/test/java/com/objects/marketbridge/order/mock/FakeMemberCouponRepository.java
@@ -1,6 +1,6 @@
 package com.objects.marketbridge.order.mock;
 
-import com.objects.marketbridge.member.domain.MemberCoupon;
+import com.objects.marketbridge.coupon.domain.MemberCoupon;
 import com.objects.marketbridge.product.infra.coupon.MemberCouponRepository;
 
 import java.util.List;
